requirements.txt
python-telegram-bot==20.8
openai==1.30.3
python-dotenv==1.0.1
